perf: try to reuse previously detected encoding when possible
